+++ hr-ai/src/main/java/com/open/hr/ai/bean/vo/AmPositionVo.java
@@ -74,7 +74,7 @@ public class AmPositionVo {
     /**
      * 渠道，platform的id
      */
-    private Integer channel;
+    private Long channel;
 
     /**
      * 招聘状态 1运行中，0暂停
